Merge: fix security error (login uri update)

See merge request s10-bigdata-recom-sub2/S10P22A108!30

Former-commit-id: 413e22d72a02d797e7036e888860bb2b6204e0f2

+++ back-end/src/main/java/com/nightletter/domain/diary/entity/DiaryTarotId.java
@@ -1,18 +1,18 @@
 package com.nightletter.domain.diary.entity;
 
-import java.io.Serializable;
-import java.util.Objects;
-
 import jakarta.persistence.Embeddable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+import java.util.Objects;
 
 @Getter
 @NoArgsConstructor
 @Embeddable
 public class DiaryTarotId implements Serializable {
-	private Long diaryId;
-	private Long tarotId;
+    private Long diaryId;
+    private Integer tarotId;
 
 	@Override
 	public boolean equals(Object obj) {
